medperf/medperf/requirements.txt
typer
PyYAML
requests
yaspin
tabulate